Changed handling of tag duels and relay duels
Unified new card functions under new_card, with an additional parameter (duelist). This parameter will be used to add any "additional" player, like a partner in a tag duel or in a relay duel, the api will automatically handle the various modes, if DUEL_RELAY_MODE is passed then the extra players will be considered like in a relay duel, otherwise the extra players will play in rotation like in a tag duel.