Another attempt for 174ec18722ad28cb13fbcc5c3b283bb036301d33
Now grant_effect_container is instead a struct containing both an ordered an unordered map, the unordered map is the primairy one, and it's the one that holds the actual values, the sorted map holds only pointers to those elements. This should avoid the crashes when iterating the map and erasing elements, while also allowing to have a sorted access to the elements.