Add lua_get<function>
Like other lua_get functions, perform the various ehecks on the parameter at the specified index and error out if the check is invalid, or return "0" if the function is called in a non erroring out way (0 is a null index, thus meaning that the paramter at that index is not a function). If the parameter is a function, returns the index itself.