Fix field::negate_chain (#147)
It was setting the statuses STATUS_LEAVE_CONFIRMED and STATUS_ACTIVATE_DISABLED only after checking if the card was in the szone, not accounting for instances where in the same chain the card left the field and got put back in on there before negating the effect. Now only set those statuese if the card is still related to the chain.