Initial work on having a way to avoid using c++ exceptions to interface with lua
We make all the lua api functions actually return a `LuaRet` type, which is noting but a variant holding either an integer, or a function object.
If the lua function returns normally, the integer would be returned, and it'll all be fine, if instead the function yields or an exception is raised, the function will return a function object which once called will call the actual lua function to perform the yield/error throwing.
This eliminates the issue of using a lua built without c++ exceptions support, by ensuring that a possible call to longjmp is performed after the defined lua function is called, so the destructors of all the temporairly allocated non trivially destructible c++ classes are properly called and no heap corruption is performed.
A current downside is that this mechanism works cleanly only if yield/lua_error are called in the main lua function body, if lua_error is called by a sub function call, this automatic handling is not possible, for now a lua_error_unsafe has been added, which will just call directly luaL_error, with the caveat that it HAS TO be called in a context where no not trivially destructible objects are present on the stack