Fix compilation error due to low PROCESSOR_VARIANT_CAP
If PROCESSOR_VARIANT_CAP was set to a value which made the subvariants have Processors::SelectPlace and Processors::SelectDisField not be part of the same variant, emplacing a new variant object caused a compile error due to ambiguity on which variant to pick.
To fix this SelectDisField is no longer listed in the processor_unit tuple so that the base class will always be taken in consideration, this prevents `emplace_back(std::in_place_type<T>,` from being used tho when not using the subvariant approach.