Use the proper methods to iterate tables when receiving them as arguments
No longer use lua_next and co, but get the table length and use lua_rawgeti to iterate all the fields.
Support for tables packed via table.pack (which have an extra field "n" to indicate the number of elements) was added as well.
This introduces a possibly breaking change due to the fact that we still use undefined behaviour, but of a different kind from the previous one.
With the previous approach, the order of elements wasn't guaranteed, and nil elements would've been skipped. With this approach, order is guaranteed, but unless the table is packed, there's no guarantee if all the elements are parsed or not (that's up to lua)